Add OpenVR binding for VIve Cosmos TODO: Firing Logic

+++ EventExtractor/WeaponState.cs
@@ -7,14 +7,14 @@ namespace EventDetectors
 {
     public class WeaponState
     {
-        private byte triggerState = 0;
+        private double triggerState = 0;
         private ushort bulletCount = 0;
         private bool isAutoFire = false;
         private DateTime lastTriggerEnter = DateTime.Now;
         private DateTime lastTriggerExit = DateTime.Now;
         
         public Publisher publisher;
-        public byte TriggerState { get => triggerState; set => triggerState = value; }
+        public double TriggerState { get => triggerState; set => triggerState = value; }
         public ushort BulletCount { get => bulletCount; set => bulletCount = value; }
         public bool IsAutoFire { get => isAutoFire; set => isAutoFire = value; }
         public DateTime LastTriggerEnter { get => lastTriggerEnter; set => lastTriggerEnter = value; }
